requirements.txt
openai>=1.0.0
anthropic>=0.54.0
google-generativeai>=0.5.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pytest>=7.0.0
pillow>=10.0.0
aiohttp>=3.9.0
asyncio-throttle>=1.0.0
tenacity>=8.0.0
pyyaml>=6.0
jinja2>=3.1.0
typing-extensions>=4.5.0
pytest-asyncio>=0.18.0
